Others/PROJECT_STUDY/AIM-Sniper-backend-main/AIM_Sniper_backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
begins==0.9
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
diskcache==5.6.3
Django==5.1.1
django-cors-headers==4.4.0
django-erd==1.0
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.9.2
pydantic_core==2.23.4
pydotplus==2.0.2
Pygments==2.18.0
PySimpleGUI==4.60.5.0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.8
requests-toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
stack-data==0.6.3
tenacity==8.5.0
tornado==6.4
traitlets==5.14.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.3
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
